+++ src/test/java/com/kilo/BitCounterTest.java
@@ -9,7 +9,7 @@ class BitCounterTest {
 
     @ParameterizedTest
     @CsvSource({"0, 0","1, 1", "2, 1", "3, 2"})
-    void test(int x, int o) {
+    void test(long x, int o) {
         BitCounter bitCounter = new BitCounter();
         int actual = bitCounter.countBits(x);
         assertEquals(o, actual);
